Test the case where there are no favorites track
