+++ app/src/androidTest/java/ch/epfl/sweng/runpharaa/CreateTrackActivityTest.java
@@ -1,4 +1,36 @@
 package ch.epfl.sweng.runpharaa;
 
+import android.net.Uri;
+import android.support.test.rule.ActivityTestRule;
+
+import com.google.android.gms.maps.model.LatLng;
+
+import org.junit.BeforeClass;
+import org.junit.Rule;
+import org.junit.Test;
+
+import java.util.HashSet;
+
+import static android.os.SystemClock.sleep;
+import static android.support.test.espresso.Espresso.onView;
+import static android.support.test.espresso.action.ViewActions.click;
+import static android.support.test.espresso.matcher.ViewMatchers.withId;
+
 public class CreateTrackActivityTest {
+
+    @Rule
+    public final ActivityTestRule<CreateTrackActivity> mActivityRule = new ActivityTestRule<>(CreateTrackActivity.class);
+
+    @BeforeClass
+    public static void initUser() {
+        User.instance = new User("FakeUser", 2000, Uri.parse(""), new HashSet<Integer>(), new HashSet<Integer>(), new LatLng(21.23, 12.112), false, "aa");
+    }
+
+    @Test
+    public void staysOnCreateTrackActivity(){
+        sleep(2000);
+        onView(withId(R.id.start_create_button)).perform(click());
+    }
+
+
 }
