+++ app/src/androidTest/java/ch/epfl/sweng/runpharaa/FavoritesFragmentTest.java
@@ -1,7 +1,10 @@
 package ch.epfl.sweng.runpharaa;
 
 import android.Manifest;
+import android.content.Context;
+import android.content.res.Resources;
 import android.os.SystemClock;
+import android.support.test.InstrumentationRegistry;
 import android.support.test.espresso.contrib.RecyclerViewActions;
 import android.support.test.espresso.intent.Intents;
 import android.support.test.rule.ActivityTestRule;
@@ -22,6 +25,7 @@ import java.util.HashSet;
 
 import ch.epfl.sweng.runpharaa.tracks.Track;
 
+import static android.os.SystemClock.sleep;
 import static android.support.test.espresso.Espresso.onData;
 import static android.support.test.espresso.Espresso.onView;
 import static android.support.test.espresso.action.ViewActions.click;
@@ -34,6 +38,7 @@ import static android.support.test.espresso.matcher.ViewMatchers.isCompletelyDis
 import static android.support.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static android.support.test.espresso.matcher.ViewMatchers.withClassName;
 import static android.support.test.espresso.matcher.ViewMatchers.withId;
+import static android.support.test.espresso.matcher.ViewMatchers.withText;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.instanceOf;
 import static org.hamcrest.CoreMatchers.is;
@@ -56,22 +61,33 @@ public class FavoritesFragmentTest {
         Intents.init();
     }
 
-    @BeforeClass
-    public static void initUser() {
+    @Before
+    public void initUser() {
         User.instance = new User("FakeUser", 2000, null, new HashSet<Integer>(), new HashSet<Integer>(), new LatLng(21.23, 12.112), false, "aa");
-        User.instance.addToFavorites(0);
-        User.instance.addToFavorites(1);
+
     }
 
 
     @Test
     public void testFavoritesAppears() {
+        User.instance.addToFavorites(0);
+        User.instance.addToFavorites(1);
         onView(withId(R.id.viewPagerId)).perform(swipeLeft());
         onView(withId(R.id.viewPagerId)).perform(swipeLeft());
-        SystemClock.sleep(5_000);
+        sleep(5_000);
         onView(allOf(withId(R.id.cardListId), isDisplayed())).perform(
                 actionOnItemAtPosition(0, click()));
         intended(hasComponent(TrackPropertiesActivity.class.getName()));
+    }
+
+    @Test
+    public void testNoFavorites(){
+        Context targetContext = InstrumentationRegistry.getInstrumentation()
+                .getTargetContext();
+        onView(withId(R.id.viewPagerId)).perform(swipeLeft());
+        onView(withId(R.id.viewPagerId)).perform(swipeLeft());
+        sleep(2000);
+        onView(allOf(withId(R.id.emptyMessage), isDisplayed())).check(matches(withText(targetContext.getResources().getString(R.string.no_favorite))));
     }
 
     @After
