+++ app/src/androidTest/java/ch/epfl/sweng/runpharaa/TrackPropertiesActivityTest.java
@@ -15,30 +15,30 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import java.util.ArrayList;
 import java.util.HashSet;
 
 import ch.epfl.sweng.runpharaa.tracks.Track;
 
+import static android.support.test.espresso.Espresso.onView;
+import static android.support.test.espresso.action.ViewActions.click;
 import static android.support.test.espresso.matcher.ViewMatchers.withId;
 import static android.support.test.espresso.matcher.ViewMatchers.withText;
 
 @RunWith(AndroidJUnit4.class)
 public class TrackPropertiesActivityTest {
 
-    @BeforeClass
-    public static void initUser() {
-        User.instance = new User("FakeUser", 2000, null, new HashSet<Integer>(), new HashSet<Integer>(), new LatLng(21.23, 12.112), false, "aa");
-    }
-
     @Rule
     public ActivityTestRule<TrackPropertiesActivity> mActivityRule =
             new ActivityTestRule<>(TrackPropertiesActivity.class, true, false);
-
     @Rule
     public GrantPermissionRule permissionRule = GrantPermissionRule.grant(
             Manifest.permission.ACCESS_FINE_LOCATION,
             Manifest.permission.ACCESS_COARSE_LOCATION);
+
+    @BeforeClass
+    public static void initUser() {
+        User.instance = new User("FakeUser", 2000, null, new HashSet<Integer>(), new HashSet<Integer>(), new LatLng(21.23, 12.112), false, "aa");
+    }
 
     @Test
     public void correctValuesDisplayedForTrack1() {
@@ -53,5 +53,19 @@ public class TrackPropertiesActivityTest {
         withId(R.id.trackLikesID).matches(withText("Likes: " + t1.getProperties().getLikes()));
     }
 
+    @Test
+    public void testLike() {
+        mActivityRule.launchActivity(null);
+        onView(withId(R.id.buttonLikeID)).perform(click());
+    }
+
+    @Test
+    public void testFavourite() {
+        mActivityRule.launchActivity(null);
+        onView(withId(R.id.buttonFavoriteID)).perform(click());
+    }
 
 }
+
+
+
